Update RiemannSums.html
Code adjusted to be more concise, more readable, and to hopefully fix a few errors:
While loop switched to for loop, so that there wouldn't be any issues with looping an incorrect number of times (and hence an incorrect number of rectangles). Now will loop exactly as many times as requested number of rectangles.